⚠ 型定義、最適化

+++ pi_yo_6/voice_list.py
@@ -1,14 +1,18 @@
 import asyncio
 import re
-from discord import ui, Interaction, SelectOption ,ButtonStyle, Embed, Guild
+from discord import Member, ui, Interaction, SelectOption ,ButtonStyle, Embed, Guild
+from typing import TYPE_CHECKING
 
 from .load_config import GC, UC
 from .embeds import EmBase
 from .synthetic_voice import SyntheticEngines
 
-open_jtalk = 'open_jtalk'
-voicevox = 'voicevox'
-coeiroink = 'coeiroink'
+if TYPE_CHECKING:
+    from .main import DataInfo
+
+open_jtalk_name = 'open_jtalk'
+voicevox_name = 'voicevox'
+coeiroink_name = 'coeiroink'
 
 
 async def embed(guild:Guild):
@@ -43,26 +47,26 @@ async def embed(guild:Guild):
 
 # Button
 class CreateView(ui.View):
-    def __init__(self, g_opts, engines:SyntheticEngines, _type=voicevox, voice='ずんだもん'):
+    def __init__(self, g_opts:dict[int, "DataInfo"], engines:SyntheticEngines, _type:str=voicevox_name, voice='ずんだもん'):
         super().__init__(timeout=None)
         self.engines = engines
         self.g_opts = g_opts
 
         # typeが機能していなかった時のため
-        if _type == open_jtalk and not engines.open_jtalk:
-            _type = None
-        elif _type == voicevox and not engines.voicevox:
-            _type = None
-        elif _type == coeiroink and not engines.coeiroink:
-            _type = None
+        if _type == open_jtalk_name and not engines.open_jtalk:
+            _type = ""
+        elif _type == voicevox_name and not engines.voicevox:
+            _type = ""
+        elif _type == coeiroink_name and not engines.coeiroink:
+            _type = ""
 
-        if _type == None:
+        if _type == "":
             if engines.open_jtalk:
-                _type = open_jtalk
+                _type = open_jtalk_name
             elif engines.voicevox:
-                _type = voicevox
+                _type = voicevox_name
             elif engines.coeiroink:
-                _type = coeiroink
+                _type = coeiroink_name
 
         self._type = _type
 
@@ -82,20 +86,20 @@ class CreateView(ui.View):
 
 
 class CreateSelect(ui.Select):
-    def __init__(self, voice, parent:'CreateView') -> None:
-        self.parent = parent
+    def __init__(self, voice, parent_view:'CreateView') -> None:
+        self.parent_view = parent_view
+        engine_name = ''
+        sp_list = []
+        if parent_view._type == open_jtalk_name and parent_view.engines.open_jtalk:
+            sp_list = parent_view.engines.open_jtalk.metas
 
-        if parent._type == open_jtalk:
-            _type = ''
-            sp_list = parent.engines.open_jtalk.metas
+        elif parent_view._type == voicevox_name and parent_view.engines.voicevox:
+            engine_name = 'VOICEVOX:'
+            sp_list = parent_view.engines.voicevox.metas
 
-        elif parent._type == voicevox:
-            _type = 'VOICEVOX:'
-            sp_list = parent.engines.voicevox.metas
-
-        elif parent._type == coeiroink:
-            _type = 'COEIROINK:'
-            sp_list = parent.engines.coeiroink.metas
+        elif parent_view._type == coeiroink_name and parent_view.engines.coeiroink:
+            engine_name = 'COEIROINK:'
+            sp_list = parent_view.engines.coeiroink.metas
 
         select_opt:list[SelectOption] = []
         check = False
@@ -105,7 +109,7 @@ class CreateSelect(ui.Select):
                 check = True
                 _default=True
             if len(select_opt) < 25:
-                select_opt.append(SelectOption(label=f'{_type}{sp["name"]}', value=sp["name"], default=_default))
+                select_opt.append(SelectOption(label=f'{engine_name}{sp["name"]}', value=sp["name"], default=_default))
         if check == False:
             select_opt[0].default = True
         super().__init__(placeholder='キュー表示', options=select_opt, row=1)
@@ -114,23 +118,26 @@ class CreateSelect(ui.Select):
     async def callback(self, interaction: Interaction):
         loop = asyncio.get_event_loop()
         loop.create_task(interaction.response.defer())
-        await interaction.message.edit(view=CreateView(g_opts=self.parent.g_opts, engines=self.parent.engines, _type=self.parent._type, voice=self.values[0]))
+        if interaction.message:
+            await interaction.message.edit(view=CreateView(g_opts=self.parent_view.g_opts, engines=self.parent_view.engines, _type=self.parent_view._type, voice=self.values[0]))
 
 
 
 class CreateSelect2(ui.Select):
-    def __init__(self, voice, parent:'CreateView') -> None:
-        if parent._type == open_jtalk:
-            _type = ''
-            sp_list = parent.engines.open_jtalk.metas
+    def __init__(self, voice, parent_view:'CreateView') -> None:
+        self.parent_view = parent_view
+        engine_name = ''
+        sp_list = []
+        if parent_view._type == open_jtalk_name and parent_view.engines.open_jtalk:
+            sp_list = parent_view.engines.open_jtalk.metas
 
-        elif parent._type == voicevox:
-            _type = 'voicevox:'
-            sp_list = parent.engines.voicevox.metas
+        elif parent_view._type == voicevox_name and parent_view.engines.voicevox:
+            engine_name = 'VOICEVOX:'
+            sp_list = parent_view.engines.voicevox.metas
 
-        elif parent._type == coeiroink:
-            _type = 'coeiroink:'
-            sp_list = parent.engines.coeiroink.metas
+        elif parent_view._type == coeiroink_name and parent_view.engines.coeiroink:
+            engine_name = 'COEIROINK:'
+            sp_list = parent_view.engines.coeiroink.metas
         
         styles = None
         for _ in sp_list:
@@ -140,10 +147,10 @@ class CreateSelect2(ui.Select):
         if not styles:
             voice = sp_list[0]['name']
             styles = sp_list[0]['styles']
-        if parent._type == open_jtalk:
+        if parent_view._type == open_jtalk_name:
             value_prefix = ''
         else:
-            value_prefix = f'{_type}{voice}_'
+            value_prefix = f'{engine_name}{voice}_'
 
         select_opt = [SelectOption(label=f"{_['name']}", value=f'{value_prefix}{_["name"]}') for _ in styles]
         select_opt[0].default = True
@@ -159,13 +166,9 @@ class CreateSelect2(ui.Select):
 
 
 class CreateButtonPlay(ui.Button):
-    def __init__(self, parent:'CreateView') -> None:
-        try:
-            from ..main import DataInfo
-            self.g_opts:dict[int, DataInfo]
-        except Exception: pass
-        self.g_opts = parent.g_opts
-        self.parent = parent
+    def __init__(self, parent_view:'CreateView') -> None:
+        self.g_opts = parent_view.g_opts
+        self.parent_view = parent_view
         super().__init__(label='Play', style=ButtonStyle.blurple, row=3)
 
 
@@ -173,11 +176,12 @@ class CreateButtonPlay(ui.Button):
         loop = asyncio.get_event_loop()
         loop.create_task(interaction.response.defer())
 
-        if Voice := self.g_opts.get(interaction.guild.id):
-            Voice = Voice.Voice
+        if not interaction.guild: return
+        if data := self.g_opts.get(interaction.guild.id):
+            Voice = data.voice
             Voice.Queue.append([interaction.id, 0])
 
-            try: source = await Voice.creat_voice(f'voice:{self.parent.select2.voice_res} テストなのだ', interaction)
+            try: source = await Voice.creat_voice(f'voice:{self.parent_view.select2.voice_res} テストなのだ', interaction.message)
             except Exception as e:                                              # Error
                 print(f"Error : 音声ファイル作成に失敗 {e}")
                 Voice.Queue.remove([interaction.id, 0])
@@ -187,32 +191,34 @@ class CreateButtonPlay(ui.Button):
             Voice.Queue[i:i+1] = [[_,1] for _ in source]
 
             # 再生されるまでループ
-            if not Voice.Vvc.is_playing():
+            if not Voice.track.has_play_data():
                 await Voice.play_loop()
 
 
 
 class CreateButtonText(ui.Button):
-    def __init__(self, parent:'CreateView') -> None:
-        self.parent = parent
+    def __init__(self, parent_view:'CreateView') -> None:
+        self.parent_view = parent_view
         super().__init__(label='読むテキスト', style=ButtonStyle.green, row=3)
 
     async def callback(self, interaction: Interaction):
         await interaction.response.defer()
+
+        if not interaction.guild: return
         _GC = GC(gid=interaction.guild.id)
         g_config = _GC.Read()
-        g_config['voice'][str(interaction.user.id)] = self.parent.select2.voice_res
+        g_config['voice'][str(interaction.user.id)] = self.parent_view.select2.voice_res
         _GC.Write(g_config)
 
 
 
 class CreateButtonSet(ui.Button):
-    def __init__(self, parent:'CreateView') -> None:
-        self.parent = parent
+    def __init__(self, parent_view:'CreateView') -> None:
+        self.parent_view = parent_view
         super().__init__(label='ボイスをセット', style=ButtonStyle.blurple, row=3)
 
     async def callback(self, interaction: Interaction):
-        new_message = EditVoiceMessage(self.parent.select2.voice_res)
+        new_message = EditVoiceMessage(self.parent_view.select2.voice_res)
         await interaction.response.send_message(embed=new_message.who_embed, view=new_message.who_view, ephemeral=True)
 
 
@@ -222,6 +228,7 @@ class CreateButtonRefresh(ui.Button):
         super().__init__(label='↺', style=ButtonStyle.blurple, row=3)
 
     async def callback(self, interaction: Interaction):
+        if not interaction.guild: return
         await interaction.response.edit_message(embed=await embed(interaction.guild))
 
 
@@ -232,46 +239,50 @@ class CreateButtonDel(ui.Button):
 
     async def callback(self, interaction: Interaction):
         await interaction.response.defer()
-        await interaction.message.delete()
+        if interaction.message:
+            await interaction.message.delete()
 
 
 
 class CreateButtonOpenJtalk(ui.Button):
-    def __init__(self, parent:'CreateView') -> None:
-        self.parent = parent
-        if parent.engines.open_jtalk:
+    def __init__(self, parent_view:'CreateView') -> None:
+        self.parent_view = parent_view
+        if parent_view.engines.open_jtalk:
             super().__init__(label='Open_Jtalk', style=ButtonStyle.green, row=0)
         else:
             super().__init__(label='Open_Jtalk', style=ButtonStyle.gray, disabled=True, row=0)
 
     async def callback(self, interaction: Interaction):
         await interaction.response.defer()
-        await interaction.message.edit(view=CreateView(g_opts=self.parent.g_opts, engines=self.parent.engines, _type=open_jtalk))
+        if interaction.message:
+            await interaction.message.edit(view=CreateView(g_opts=self.parent_view.g_opts, engines=self.parent_view.engines, _type=open_jtalk_name))
 
 
 class CreateButtonVoicevox(ui.Button):
-    def __init__(self, parent:'CreateView') -> None:
-        self.parent = parent
-        if parent.engines.voicevox:
+    def __init__(self, parent_view:'CreateView') -> None:
+        self.parent_view = parent_view
+        if parent_view.engines.voicevox:
             super().__init__(label='VoiceVox', style=ButtonStyle.green, row=0)
         else:
             super().__init__(label='VoiceVox', style=ButtonStyle.gray, disabled=True, row=0)
 
     async def callback(self, interaction: Interaction):
         await interaction.response.defer()
-        await interaction.message.edit(view=CreateView(g_opts=self.parent.g_opts, engines=self.parent.engines, _type=voicevox))
+        if interaction.message:
+            await interaction.message.edit(view=CreateView(g_opts=self.parent_view.g_opts, engines=self.parent_view.engines, _type=voicevox_name))
 
 class CreateButtonCoeiroink(ui.Button):
-    def __init__(self, parent:'CreateView') -> None:
-        self.parent = parent
-        if parent.engines.coeiroink:
+    def __init__(self, parent_view:'CreateView') -> None:
+        self.parent_view = parent_view
+        if parent_view.engines.coeiroink:
             super().__init__(label='Coeiroink', style=ButtonStyle.green, row=0)
         else:
             super().__init__(label='Coeiroink', style=ButtonStyle.gray, disabled=True, row=0)
 
     async def callback(self, interaction: Interaction):
         await interaction.response.defer()
-        await interaction.message.edit(view=CreateView(g_opts=self.parent.g_opts, engines=self.parent.engines, _type=coeiroink))
+        if interaction.message:
+            await interaction.message.edit(view=CreateView(g_opts=self.parent_view.g_opts, engines=self.parent_view.engines, _type=coeiroink_name))
 
 
 
@@ -286,7 +297,10 @@ class EditVoiceMessage:
 
             @ui.button(label='自分', style=ButtonStyle.green)
             async def my_voice(self, interaction:Interaction, button:ui.Button):
-                gid = interaction.guild_id
+                if not interaction.guild: return 
+                if not isinstance(interaction.user, Member): return
+
+                gid = interaction.guild.id
                 _GC = GC(gid)
                 g_config = _GC.Read()
                 
@@ -300,7 +314,10 @@ class EditVoiceMessage:
 
             @ui.button(label='他人', style=ButtonStyle.green)
             async def another_voice(self, interaction:Interaction, button:ui.Button):
-                gid = interaction.guild_id
+                if not interaction.guild: return 
+                if not isinstance(interaction.user, Member): return
+                
+                gid = interaction.guild.id
                 _GC = GC(gid)
                 g_config = _GC.Read()
                 
@@ -314,7 +331,10 @@ class EditVoiceMessage:
 
             @ui.button(label='サーバーの初期ボイス', style=ButtonStyle.green)
             async def server_voice(self, interaction:Interaction, button:ui.Button):
-                gid = interaction.guild_id
+                if not interaction.guild: return 
+                if not isinstance(interaction.user, Member): return
+                
+                gid = interaction.guild.id
                 _GC = GC(gid)
                 g_config = _GC.Read()
